fix: move clustersResource into right directory
https://gravitee.atlassian.net/browse/APIM-10045
